# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Laravel API",
        "description": "API documentation for my Laravel application.",
        "version": "1.0"
    },
    "paths": {
...
        "/api/preferences": {
            "get": {
                "tags": [
                    "Preferences"
                ],
                "summary": "Get user preferences",
                "operationId": "86d82a29efa9769882266a99f89e8f6d",
                "responses": {
                    "200": {
                        "description": "User preferences retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Preference"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Preferences not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Preferences"
                ],
                "summary": "Set user preferences",
                "operationId": "a00b29a1fbbd8572ac8f659146ade2da",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sources": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "NewsAPI",
                                            "The Guardian"
                                        ]
                                    },
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "sports",
                                            "general"
                                        ]
                                    },
                                    "authors": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "John Doe",
                                            "Jane Smith"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Preferences saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Preferences saved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Preference"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/personalized-feed": {
            "get": {
                "tags": [
                    "Preferences"
                ],
                "summary": "Get personalized news feed based on user preferences",
                "operationId": "69a4f37dd10424fbca9f28f187d71626",
                "responses": {
                    "200": {
                        "description": "Personalized news feed retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Article"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No articles found based on preferences"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Article": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Sample Article Title"
                    },
                    "content": {
                        "type": "string",
                        "example": "Content of the article goes here."
                    },
                    "author": {
                        "type": "string",
                        "example": "Author Name"
                    },
                    "source": {
                        "type": "string",
                        "example": "Source Name"
                    },
                    "category": {
                        "type": "string",
                        "example": "Category Name"
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-11T12:34:56Z"
                    }
                },
                "type": "object"
            },
            "Preference": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "sources": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "authors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-11T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-11T12:34:56Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}